requirements.txt
pika~=1.3.2
torch
torchvision
numpy
requests~=2.32.3
pyyaml
opencv-python
ultralytics
